statistics

+++ diffractio/utils_drawing.py
@@ -306,7 +306,7 @@ def extract_image_from_video(nombre_video: str | None = None, num_frame: str = "
     os.system(texto)
 
 
-def normalize_draw(u, logarithm: float | bool = False, normalize: bool = False, cut_value: int | None = None):
+def normalize_draw(u, logarithm: float | bool = False, normalize: bool = False, cut_value: float | None = None):
     """
     Gets a field and changes its caracteristics for drawing
